+++ salt/modules/saltutil.py
@@ -547,6 +547,44 @@ def sync_proxymodules(saltenv=None, refresh=False, extmod_whitelist=None, extmod
     return ret
 
 
+def sync_matchers(saltenv=None, refresh=False, extmod_whitelist=None, extmod_blacklist=None):
+    '''
+    .. versionadded:: Flourine
+
+    Sync engine modules from ``salt://_matchers`` to the minion
+
+    saltenv
+        The fileserver environment from which to sync. To sync from more than
+        one environment, pass a comma-separated list.
+
+        If not passed, then all environments configured in the :ref:`top files
+        <states-top>` will be checked for engines to sync. If no top files are
+        found, then the ``base`` environment will be synced.
+
+    refresh : True
+        If ``True``, refresh the available execution modules on the minion.
+        This refresh will be performed even if no new matcher modules are synced.
+        Set to ``False`` to prevent this refresh.
+
+    extmod_whitelist : None
+        comma-separated list of modules to sync
+
+    extmod_blacklist : None
+        comma-separated list of modules to blacklist based on type
+
+    CLI Examples:
+
+    .. code-block:: bash
+
+        salt '*' saltutil.sync_matchers
+        salt '*' saltutil.sync_matchers saltenv=base,dev
+    '''
+    ret = _sync('matchers', saltenv, extmod_whitelist, extmod_blacklist)
+    if refresh:
+        refresh_modules()
+    return ret
+
+
 def sync_engines(saltenv=None, refresh=False, extmod_whitelist=None, extmod_blacklist=None):
     '''
     .. versionadded:: 2016.3.0
@@ -943,6 +981,7 @@ def sync_all(saltenv=None, refresh=True, extmod_whitelist=None, extmod_blacklist
     ret['engines'] = sync_engines(saltenv, False, extmod_whitelist, extmod_blacklist)
     ret['thorium'] = sync_thorium(saltenv, False, extmod_whitelist, extmod_blacklist)
     ret['serializers'] = sync_serializers(saltenv, False, extmod_whitelist, extmod_blacklist)
+    ret['matchers'] = sync_matchers(saltenv, False, extmod_whitelist, extmod_blacklist)
     if __opts__['file_client'] == 'local':
         ret['pillar'] = sync_pillar(saltenv, False, extmod_whitelist, extmod_blacklist)
     if refresh:
@@ -965,6 +1004,24 @@ def refresh_beacons():
         ret = __salt__['event.fire']({}, 'beacons_refresh')
     except KeyError:
         log.error('Event module not available. Module refresh failed.')
+        ret = False  # Effectively a no-op, since we can't really return without an event system
+    return ret
+
+
+def refresh_matchers():
+    '''
+    Signal the minion to refresh its matchers.
+
+    CLI Example:
+
+    .. code-block:: bash
+
+        salt '*' saltutil.refresh_matchers
+    '''
+    try:
+        ret = __salt__['event.fire']({}, 'matchers_refresh')
+    except KeyError:
+        log.error('Event module not available. Matcher refresh failed.')
         ret = False  # Effectively a no-op, since we can't really return without an event system
     return ret
